Add CSS support to the webpack configuration
While we're building the site, there are many speed and performance
advantages to have our CSS handled by our JavaScript.

For example, in the future we'll setup webpack's development server so
that when we save a change to our CSS, we don't even need to refresh in
the web browser.