Partially style the title and the subtitle
By applying the BEM philosophy, we can make the relationship between our
CSS and HTML crystal clear. We can also combine it with the power we get
when using the postcss-nested package to make our code super organized.

In summary:

 - You have to give an element a class name related to the block it
   belongs to.

 - You have to nest the elements' styling inside the block styling
   (replace the route rules selector for an & to make webpack
   generate the exact name of the class we have given to an element).

Doing so creates an one-to-one relationship between our CSS and HTML, so
we don't need to rely on descendant selectors and the cascade.